Parameterized the extraInfo in the onion header.
The paper states that the overhead is 1 group element + (next hop
address + header mac)*number of hops + 1 payload MAC. Previously the
payload MAC was also added to the onion, which resulted in strange
lengths for the header. Now it is parameterized and we can remove it at
will by setting extraInfo=0.